+++ test.html
@@ -1,0 +1,16 @@
+<!DOCTYPE HTML>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <!--
+      The above 2 meta tags *must* come first in the <head>
+      to consistently ensure proper document rendering.
+      Any other head element should come *after* these tags.
+     -->
+    <title>Whatsnext Browser Test</title>
+  </head>
+  <body>
+    <script type="module" src="./test.js"></script>
+  </body>
+</html>
